sciencegenome
all nucletoide codes bug fixed and now putting the final protein ones and prepairing for the final release. Graph kmer finished today and now writing the graph fourier transform.

+++ src/astruct.rs
@@ -15,7 +15,7 @@ pub struct AlignmentStat {
     pub basec: usize,
     pub basen: usize,
     pub baseabsent: usize,
-    pub gccontent: usize,
+    pub gccontent: f32,
 }
 
 #[derive(Debug, Clone, PartialEq, PartialOrd)]
